Fixing signed/unsigned comparison in KdTreeNF. Updating utilities for nearest_neighbor to latest version (only change is to fix a broken test).
git-svn-id: svn+ssh://parasol-svn.cs.tamu.edu/research/parasol-svn/svnrepository/pmpl/branches/simulator-refactoring@5971 5e2c2758-810a-0410-8eb3-8e4e92d4004e

+++ src/MPLibrary/NeighborhoodFinders/KdTreeNF.h
@@ -239,7 +239,7 @@ FindNeighbors(RoadmapType* _rmp,
   // Insert scaled query (copy of original CFG).
   vector<double> queryCfg = _cfg.GetData();
   if(m_useScaling)
-    for(int i = 0; i < _cfg.PosDOF(); ++i)
+    for(size_t i = 0; i < _cfg.PosDOF(); ++i)
       queryCfg[i] /= m_maxBBXRange;
 
   PointD query(dim, queryCfg.begin(), queryCfg.end());
@@ -295,7 +295,7 @@ UpdateInternalModel(RoadmapType* _rmp, InputIterator _first,
       vector<double> cfgData = cfg.GetData();
 
       if(m_useScaling)
-        for(int i = 0; i < cfg.PosDOF(); ++i)
+        for(size_t i = 0; i < cfg.PosDOF(); ++i)
           cfgData[i] /= m_maxBBXRange;
 
       m_trees[_rmp].insert(PointD(vid, cfg.DOF(), cfgData.begin(),
